Restructure project file tree to implement taskrunner service Put the main server files into its own folder (along with package.json) Create a folder for the taskrunner and init its own package.json Reconfigure Docker to add taskrunner
