+++ server/controllers/courses/courses_list_useful.js
