+++ server/controllers/instructors/instructors_by_id.js
@@ -1,0 +1,4 @@
+// Retrieve an instructor by id
+instructors_by_id = (req,res) => {};
+
+module.exports = instructors_by_id;
