+++ server/controllers/util/courses/courses_get_by_id.js
@@ -1,0 +1,6 @@
+// Retrieve a course by id and return as JSON
+courses_get_by_id = (id, db) => {
+
+};
+
+module.exports = courses_get_by_id;
