+++ server/models/User.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
-const faculty_program_validator = require('../utils/database/validators/validate_faculty_program');
+const faculty_program_validator = require('../../utils/database/validators/validate_faculty_program');
 const TimestampType = require('./types/Timestamp');
 const OverallRatingType = require('./types/OverallRating');
