+++ server/models/types/CompactCourseRating.js
