+++ server/models/types/CompactRatingAuthor.js
