+++ server/server.js
@@ -3,7 +3,7 @@ const app = express();
 const PORT = 5000;
 
 const mongoose = require('mongoose');
-mongoose.connect('mongodb://db-mongo:27017/uwcourseapi', {useNewUrlParser: true});
+mongoose.connect('mongodb://mongodb:27017/uwcourseapi', {useNewUrlParser: true});
 const db = mongoose.connection;
 
 db.on('error', console.error.bind(console, 'Connection error: unable to connect to uwcourseapi database'));
