+++ taskrunner/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "uwcourse-taskrunner",
+  "version": "1.0.0",
+  "description": "",
+  "main": "start.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Duc Thanh Nguyen",
+  "license": "ISC",
+  "dependencies": {
+    "mongoose": "^5.4.16",
+    "node-cron": "^2.0.3",
+    "uwaterloo-api": "^1.0.2"
+  }
+}
