+++ taskrunner/start.js
@@ -1,0 +1,3 @@
+console.log('Taskrunner started');
+
+setInterval(function(){ console.log("Hello"); }, 3000);
