--- uwopendata/.gitignore
@@ -1,2 +1,0 @@
-# Hide the API keys
-config.js
